requirements.txt
requests==2.10.0
python-dotenv==0.5.1
python-telegram-bot==4.2.0
redis==2.10.5